+++ app/src/main/java/com/orange/ods/app/ui/modules/Module.kt
@@ -14,7 +14,6 @@ package com.orange.ods.app.ui.modules
 
 import androidx.annotation.DrawableRes
 import androidx.annotation.StringRes
-import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import com.orange.ods.app.R
 
@@ -25,7 +24,6 @@ sealed class Module(
     @DrawableRes val imageRes: Int,
     @StringRes val descriptionRes: Int,
     val route: String,
-    val demoScreen: @Composable (upPress: () -> Unit) -> Unit,
     val imageAlignment: Alignment = Alignment.Center,
 ) {
     val id: Long = Module::class.sealedSubclasses.indexOf(this::class).toLong()
@@ -34,7 +32,13 @@ sealed class Module(
         R.string.module_about,
         R.drawable.il_about,
         R.string.module_about_description,
-        route = ModuleDemoDestinations.AboutCustomizationRoute,
-        demoScreen = { _ -> },
+        route = ModulesNavigation.AboutCustomizationRoute
+    )
+
+    data object EmptyState : Module(
+        R.string.module_emptyState_title,
+        R.drawable.il_empty_state,
+        R.string.module_emptyState_description,
+        route = ModulesNavigation.EmptyStateCustomizationRoute
     )
 }